Transfer DffPlot into single-file component
