--- src/components/ROIPlot/template.html
@@ -1,3 +1,0 @@
-<div style="float:left;">
-  <canvas v-bind:width="width" v-bind:height="height"></canvas>
-</div>
